Improved bombardment behaviour. Added lethal bombardment behaviour. Split MapUnit class to multiple files. Minor fixes/refactors.
