+++ C7/Lua/civ3/ruleset.json
@@ -2798,6 +2798,10 @@
         "Maya"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Land"
       ],
@@ -3953,6 +3957,9 @@
         "Jet Fighter"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -4026,6 +4033,10 @@
         "Maya"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -4175,6 +4186,9 @@
         "Maya"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -4249,6 +4263,10 @@
         "Maya"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -4323,6 +4341,10 @@
         "Maya"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -5239,6 +5261,10 @@
         "America"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Air"
       ],
@@ -5690,6 +5716,10 @@
         "Artillery"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalLandBombardment",
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Land"
       ],
@@ -7425,6 +7455,9 @@
         "Caravel"
       ],
       "unproducible": false,
+      "flags": [
+        "lethalSeaBombardment"
+      ],
       "categories": [
         "Sea"
       ],
