+++ C7Engine/C7GameData/MapUnit.cs
@@ -1,20 +1,18 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+using C7Engine;
+using Serilog;
 using static C7GameData.PlayerRelationship;
-using static C7GameData.Tile;
 
 namespace C7GameData {
-	using Serilog;
-	using System.Collections.Generic;
-	using System.Text.Json.Serialization;
-	using System.Linq;
-	using C7Engine;
-	using AIData;
-	using System;
-	using System.Threading.Tasks;
-
 	/**
 	 * A unit on the map.  Not to be confused with a unit prototype.
 	 **/
-	public class MapUnit {
+	public partial class MapUnit {
+		private static ILogger log = Log.ForContext<MapUnit>();
 		public ID id { get; internal set; }
 		public string name { get; internal set; }
 		public Civilization nationality { get; set; }
@@ -65,8 +63,10 @@ namespace C7GameData {
 
 		internal MapUnit() { }
 
+		public static MapUnit NONE = new MapUnit(ID.None("unit"));
+
 		public static bool IsMapUnitValid(MapUnit mapUnit) {
-			return mapUnit != null && mapUnit != MapUnit.NONE;
+			return mapUnit != null && mapUnit != NONE;
 		}
 
 		public bool IsBusy() {
@@ -116,7 +116,7 @@ namespace C7GameData {
 		}
 
 		public override string ToString() {
-			if (this != MapUnit.NONE) {
+			if (this != NONE) {
 				return $"{this.owner} {this.GetDisplayName()} at [{this.location.XCoordinate}, {this.location.YCoordinate}] " +
 					   $"with {this.movementPoints.getMixedNumber()} MP and {this.hitPointsRemaining} HP, id = {id}";
 			} else {
@@ -205,10 +205,6 @@ namespace C7GameData {
 			}
 		}
 
-		public static MapUnit NONE = new MapUnit(ID.None("unit"));
-
-		private static ILogger log = Log.ForContext<MapUnit>();
-
 		private const int JOB_PROGRESS_WORKER = 2;
 		private const int JOB_PROGRESS_SLAVE = 1;
 
@@ -255,16 +251,6 @@ namespace C7GameData {
 			if (!canMoveFreely) return false;
 
 			return true;
-		}
-
-		public void fortify() {
-			ResetFacingDirection();
-			isFortified = true;
-			animate(MapUnit.AnimatedAction.FORTIFY);
-		}
-
-		public void wake() {
-			isFortified = false;
 		}
 
 		public void ResetFacingDirection() {
@@ -348,7 +334,7 @@ namespace C7GameData {
 				promotionChance *= 2;
 			if (GameData.rng.NextDouble() < promotionChance) {
 				Promote();
-				animate(MapUnit.AnimatedAction.VICTORY);
+				animate(AnimatedAction.VICTORY);
 			}
 		}
 
@@ -373,334 +359,6 @@ namespace C7GameData {
 			return GetAttackAnimationDirection(attackDirection.reversed());
 		}
 
-		public async Task<CombatResult> fight(MapUnit defender) {
-			var attacker = this;
-
-			// Set combat animation facing. We'll restore the defender's original facing direction at the end of the battle.
-			TileDirection attackerAttackDirection = attacker.location.directionTo(defender.location);
-			TileDirection defenderDefenseDirection = attackerAttackDirection.reversed();
-			var defenderOriginalDirection = defender.facingDirection;
-			attacker.facingDirection = attacker.GetAttackAnimationDirection(attackerAttackDirection);
-			defender.facingDirection = defender.GetAttackAnimationDirection(defenderDefenseDirection);
-
-			IEnumerable<StrengthBonus> attackBonuses  = attacker.ListStrengthBonusesVersus(defender, CombatRole.Attack , attackerAttackDirection),
-								   defenseBonuses = defender.ListStrengthBonusesVersus(attacker, CombatRole.Defense, attackerAttackDirection);
-
-			double attackerStrength = attacker.unitType.attack  * StrengthBonus.ListToMultiplier(attackBonuses),
-			   defenderStrength = defender.unitType.defense * StrengthBonus.ListToMultiplier(defenseBonuses);
-
-			log.Information($"Combat log: {attacker} ({attackerStrength}) attacking {defender} ({defenderStrength})");
-			log.Information($"\tAttacker: {attacker.unitType.name}, base strength {attacker.unitType.BaseStrength(CombatRole.Attack)}");
-			foreach (StrengthBonus bonus in attackBonuses)
-				log.Information($"\t\t+{100.0 * bonus.amount}%\t{bonus.description}");
-			log.Information($"\tDefender: {defender.unitType.name}, base strength {defender.unitType.BaseStrength(CombatRole.Defense)}");
-			foreach (StrengthBonus bonus in defenseBonuses)
-				log.Information($"\t\t+{100.0 * bonus.amount}%\t{bonus.description}");
-
-			CombatResult result = CombatResult.Impossible;
-
-			double attackerOdds = attackerStrength / (attackerStrength + defenderStrength);
-			if (Double.IsNaN(attackerOdds))
-				return result;
-
-			// Defensive bombard
-			MapUnit defensiveBombarder = MapUnit.NONE;
-			double defensiveBombarderStrength = 0.0;
-			foreach (MapUnit candidate in defender.location.unitsOnTile.Where(u => u != defender && !u.owner.IsAtPeaceWith(attacker.owner) && u.defensiveBombardsRemaining > 0)) {
-				double strength = candidate.StrengthVersus(attacker, CombatRole.DefensiveBombard, defenderDefenseDirection);
-				if (strength > defensiveBombarderStrength) {
-					defensiveBombarder = candidate;
-					defensiveBombarderStrength = strength;
-				}
-			}
-			// In the original game, defensive bombard does not trigger against attackers with 1 HP. See:
-			// https://github.com/C7-Game/Prototype/pull/250#discussion_r893051111
-			if (defensiveBombarder != MapUnit.NONE && attacker.hitPointsRemaining > 1) {
-				var dBOriginalDirection = defensiveBombarder.facingDirection;
-				TileDirection defensiveBombardDirection = defenderDefenseDirection;
-				defensiveBombarder.facingDirection = defensiveBombarder.GetAttackAnimationDirection(defensiveBombardDirection);
-
-				await defensiveBombarder.animateAsync(MapUnit.AnimatedAction.ATTACK1);
-
-				// dADB = defense Against Defensive Bombard
-				double dADB = attacker.StrengthVersus(defensiveBombarder, CombatRole.DefensiveBombardDefense, defensiveBombardDirection);
-				if (GameData.rng.NextDouble() < defensiveBombarderStrength / (defensiveBombarderStrength + dADB))
-					attacker.hitPointsRemaining -= 1;
-
-				defensiveBombarder.defensiveBombardsRemaining -= 1;
-				defensiveBombarder.facingDirection = dBOriginalDirection;
-			}
-
-			bool defenderEligibleToRetreat = defender.hitPointsRemaining > 1 && ! defender.location.HasCity();
-
-			// Do combat rounds
-			while (true) {
-				defender.animate(MapUnit.AnimatedAction.ATTACK1);
-				await attacker.animateAsync(MapUnit.AnimatedAction.ATTACK1);
-				if (GameData.rng.NextDouble() < attackerOdds) {
-					if (defenderEligibleToRetreat &&
-						defender.hitPointsRemaining == 1 &&
-						GameData.rng.NextDouble() < defender.RetreatChance(attacker, false)) {
-						// TODO: Defender retreat behavior requires some more work. There's an issue for it here:
-						// https://github.com/C7-Game/Prototype/issues/274
-						Tile retreatDestination = defender.location.neighbors[attackerAttackDirection];
-						if ((retreatDestination != Tile.NONE) && defender.CanEnter(retreatDestination)) {
-							await defender.move(attackerAttackDirection, true);
-							result = CombatResult.DefenderRetreated;
-							break;
-						}
-					}
-					defender.hitPointsRemaining -= 1;
-					if (defender.hitPointsRemaining <= 0) {
-						result = CombatResult.DefenderKilled;
-						break;
-					}
-				} else {
-					if (attacker.hitPointsRemaining == 1 &&
-						GameData.rng.NextDouble() < attacker.RetreatChance(defender, true)) {
-						result = CombatResult.AttackerRetreated;
-						break;
-					}
-					attacker.hitPointsRemaining -= 1;
-					if (attacker.hitPointsRemaining <= 0) {
-						result = CombatResult.AttackerKilled;
-						break;
-					}
-				}
-			}
-
-			if ((result == CombatResult.AttackerKilled) || (result == CombatResult.DefenderKilled)) {
-				var (dead, alive) = (result == CombatResult.AttackerKilled) ? (attacker, defender) : (defender, attacker);
-				alive.RollToPromote(dead);
-				await dead.animateAsync(MapUnit.AnimatedAction.DEATH);
-				dead.RemoveFromPlay();
-			}
-
-			if (result.DefenderWon())
-				defender.facingDirection = defenderOriginalDirection;
-
-			return result;
-		}
-
-		public bool canBombard() {
-			return unitType.actions.Contains(UnitAction.Bombard);
-		}
-
-		public bool canBombardTile(Tile tile) {
-			if (unitType.bombard == 0)
-				return false;
-
-			if (tile.HasImprovements)
-				return true;
-
-			MapUnit target = tile.FindTopDefender(this);
-			if (target.owner == owner)
-				return false;
-
-			if (target != MapUnit.NONE)
-				return true;
-
-			if (tile.HasCity() && tile.cityAtTile.owner != owner)
-				return true;
-
-			return false;
-		}
-
-
-		public async Task bombard(Tile tile) {
-			// Could check canBombardTile(..) again, but no need really
-
-			MapUnit target = tile.FindTopDefender(this);
-
-			var hasTargetUnit = target != MapUnit.NONE && target.owner != owner;
-			var hasForeignCity = tile.HasCity() && tile.cityAtTile.owner != owner;
-			var hasCityWalls = hasForeignCity && tile.cityAtTile.GetBuildings().Any(b => b.building.providesWalls);
-			var hasTileImprovements = tile.HasImprovements;
-
-			if (!(hasTargetUnit || hasTileImprovements || hasForeignCity))
-				return; // Nothing to bombard
-
-			var unitOriginalOrientation = facingDirection;
-			facingDirection = location.directionTo(tile);
-
-			if (hasCityWalls)
-				await bombardCityWalls(tile);
-			else if (hasTargetUnit)
-				await bombardUnits(tile, target);
-			else if (hasForeignCity)
-				await bombardCity(tile);
-			else
-				await bombardTileImprovements(tile);
-
-			facingDirection = unitOriginalOrientation;
-		}
-
-		private async Task bombardUnits(Tile tile, MapUnit target) {
-			// TODO: Make configurable
-
-			var hitCount = 0;
-
-			foreach (var fire in Enumerable.Range(0, unitType.rateOfFire)) {
-				// TODO: Figure out the bombard defense that walls grant.
-				double bombardStrength  = StrengthVersus(target, CombatRole.Bombard, facingDirection);
-				double defenderStrength = target.StrengthVersus(this, CombatRole.BombardDefense, facingDirection);
-				double attackerOdds = bombardStrength / (bombardStrength + defenderStrength);
-				if (Double.IsNaN(attackerOdds))
-					return;
-
-				// TODO: Lethal/non-lethal bombardment
-				var isPotentiallyLethal = target.hitPointsRemaining == 1;
-
-				await animateAsync(MapUnit.AnimatedAction.ATTACK1);
-				movementPoints.onUnitMove(1);
-				if (GameData.rng.NextDouble() < attackerOdds && !isPotentiallyLethal) {
-					hitCount += 1;
-					target.hitPointsRemaining -= 1;
-					await tile.AnimateAsync(AnimatedEffect.Hit3);
-				} else
-					await tile.AnimateAsync(AnimatedEffect.Miss);
-
-				if (target.hitPointsRemaining <= 0) {
-					RollToPromote(target);
-					await target.animateAsync(MapUnit.AnimatedAction.DEATH);
-					target.RemoveFromPlay();
-					break; // Target destroyed, skip remaining fire -- TODO: Re-target?
-				}
-			}
-
-			if (owner.isHuman) {
-				if (hitCount > 0)
-					new MsgShowTemporaryPopup($"Artillery bombardment successful! Enemy units injured.", tile).send();
-				else
-					new MsgShowTemporaryPopup($"Artillery bombardment failed.", tile).send();
-			}
-		}
-
-		private async Task bombardCityWalls(Tile tile) {
-			// CF Civilopedia: City walls have a land bombardment defense of 8
-			// CF Civilopedia: Coastal defences have a land bombardment defense of 8
-			// Anecdotal: "City walls are hit first."
-			// TODO: Make configurable
-
-			const int wallDefence = 8;
-
-			var hitCount = 0;
-
-			var walls = tile.cityAtTile.GetBuildings().First(b => b.building.providesWalls);
-
-			foreach (var fire in Enumerable.Range(0, unitType.rateOfFire)) {
-				double bombardStrength  = StrengthVersus(null, CombatRole.Bombard, facingDirection);
-				double defenderStrength = wallDefence;
-				double attackerOdds = bombardStrength / (bombardStrength + defenderStrength);
-				if (Double.IsNaN(attackerOdds))
-					return;
-
-				await RunAnimatedBombard(tile, attackerOdds, () => {
-					hitCount += 1;
-					tile.cityAtTile.RemoveBuilding(walls);
-				});
-			}
-
-			if (owner.isHuman) {
-				if (hitCount > 0)
-					new MsgShowTemporaryPopup($"Artillery bombardment successful! Walls destroyed.", tile).send();
-				else
-					new MsgShowTemporaryPopup($"Artillery bombardment failed.", tile).send();
-			}
-		}
-
-		private async Task RunAnimatedBombard(Tile tile, double attackerOdds, Action callback) {
-			await animateAsync(MapUnit.AnimatedAction.ATTACK1);
-			movementPoints.onUnitMove(1);
-			if (GameData.rng.NextDouble() < attackerOdds) {
-				await tile.AnimateAsync(AnimatedEffect.Hit3);
-				callback();
-			} else
-				await tile.AnimateAsync(AnimatedEffect.Miss);
-		}
-
-		private async Task bombardCity(Tile tile) {
-			// Anecdotal: If there are no units left to hit, then citizens or buildings are hit, apparently with same probability.
-			// Anecdotal: "buildings (if I remember correctly) have a defense value of 16"
-			// Anecdotal: It seems population is killed off more quickly than buildings.
-			// TODO: Make configurable
-
-			const int buildingDefence = 16;
-			const int populationDefence = 12;
-			const float buildingOrPopulationOdds = 0.5f;
-
-			var targetBuildings = GameData.rng.NextDouble() <= buildingOrPopulationOdds;
-			var defence = targetBuildings ? buildingDefence : populationDefence;
-			var destroyMsg = targetBuildings ? "Destroyed city population." : "Destroyed a building.";
-			Action remover = targetBuildings
-				? () =>
-				{
-					var building = tile.cityAtTile.GetBuildings()
-						.OrderBy(x => GameData.rng.Next()).FirstOrDefault();
-					tile.cityAtTile.RemoveBuilding(building);
-				}
-			: () =>
-				{
-					tile.cityAtTile.RemoveRandomCitizen();
-				};
-
-			var hitCount = 0;
-
-			foreach (var fire in Enumerable.Range(0, unitType.rateOfFire)) {
-				double bombardStrength  = StrengthVersus(null, CombatRole.Bombard, facingDirection);
-				double defenderStrength = defence;
-				double attackerOdds = bombardStrength / (bombardStrength + defenderStrength);
-				if (Double.IsNaN(attackerOdds))
-					return;
-
-				await RunAnimatedBombard(tile, attackerOdds, () => {
-					hitCount += 1;
-					remover();
-				});
-			}
-
-			if (owner.isHuman) {
-				if (hitCount > 0)
-					new MsgShowTemporaryPopup($"Artillery bombardment successful! {destroyMsg}", tile).send();
-				else
-					new MsgShowTemporaryPopup($"Artillery bombardment failed.", tile).send();
-			}
-		}
-
-		private async Task bombardTileImprovements(Tile tile) {
-			// Anecdotal: "arty seems to wipe out improvement on 75% or more of the shots"
-			// ==> Artillery.bombard : 12 --> TileImprovement.Defense : 3
-			// TODO: Make configurable
-
-			const int tileImprovementDefence = 3;
-
-			var hitCount = 0;
-
-			var improvement = tile.overlays.GetImprovements()
-				.OrderBy(x => GameData.rng.Next()).FirstOrDefault();
-
-			foreach (var fire in Enumerable.Range(0, unitType.rateOfFire)) {
-				double bombardStrength  = StrengthVersus(null, CombatRole.Bombard, facingDirection);
-				double defenderStrength = tileImprovementDefence;
-				double attackerOdds = bombardStrength / (bombardStrength + defenderStrength);
-				if (Double.IsNaN(attackerOdds))
-					return;
-
-				await RunAnimatedBombard(tile, attackerOdds, () => {
-					hitCount += 1;
-					tile.overlays.Remove(improvement);
-					// TODO: Re-target?
-				});
-			}
-
-			if (owner.isHuman) {
-				if (hitCount > 0)
-					new MsgShowTemporaryPopup($"Artillery bombardment successful! Destroyed {improvement?.key}.", tile).send();
-				else
-					new MsgShowTemporaryPopup($"Artillery bombardment failed.", tile).send();
-			}
-		}
-
 		public int HealRateAt(Tile location) {
 			GameData gD = EngineStorage.gameData;
 			City city = location.cityAtTile;
@@ -711,73 +369,6 @@ namespace C7GameData {
 				return 0;
 			return gD.healRateInNeutralField;
 			// TODO: Consider friendly/neutral/enemy territory once that's implemented, barracks, the Red Cross
-		}
-
-		public void OnBeginTurn(bool skipTurn = false) {
-			int maxMP = unitType.movement;
-			if (movementPoints.remaining >= maxMP && !skipTurn) {
-				int maxHP = maxHitPoints;
-				if (hitPointsRemaining < maxHP)
-					hitPointsRemaining += HealRateAt(location);
-				if (hitPointsRemaining > maxHP)
-					hitPointsRemaining = maxHP;
-			}
-
-			if (skipTurn) {
-				movementPoints.skipTurn();
-			} else {
-				movementPoints.reset(maxMP);
-			}
-
-			defensiveBombardsRemaining = 1;
-		}
-
-		public void OnEnterTile(Tile tile) {
-			//Add to player knowledge of tiles
-			owner.tileKnowledge.AddTilesToKnown(tile);
-
-			// Disperse barb camp
-			if (tile.hasBarbarianCamp && !owner.isBarbarians) {
-				EngineStorage.gameData.map.barbarianCamps.Remove(tile);
-				tile.hasBarbarianCamp = false;
-				animate(MapUnit.AnimatedAction.VICTORY);
-
-				// TODO: make this configurable
-				owner.gold += 25;
-				if (owner.isHuman) {
-					new MsgShowMilitaryAdvisorPopup($"We cleared a barbarian encampment and earned 25 gold!", happy: true).send();
-				}
-			}
-
-			// Destroy the enemy city on the tile unless we're the barbarians,
-			// in which case we'll just take some gold.
-			if (tile.HasCity() && !owner.IsAtPeaceWith(tile.cityAtTile.owner)) {
-				if (owner.isBarbarians) {
-					// TODO: Add rules for how much gold is taken.
-					int goldTaken = tile.cityAtTile.owner.gold / 4;
-					tile.cityAtTile.owner.gold -= goldTaken;
-					this.RemoveFromPlay();
-					if (tile.cityAtTile.owner.isHuman) {
-						new MsgShowMilitaryAdvisorPopup($"Barbarians have stolen {goldTaken} gold from our cities!\nWe need a stronger military.", happy: false).send();
-					}
-				} else {
-					CityInteractions.DestroyCity(tile.XCoordinate, tile.YCoordinate);
-				}
-			}
-
-			// Check to see if we've discovered a new civ.
-			//
-			// TODO: this should really be based on interactions with our "visible"
-			// tiles. Also civ3 only counts border-based discovery from rank 1
-			// tiles, not rank 2+.
-			foreach (Tile t in tile.neighbors.Values) {
-				if (t.unitsOnTile.Count > 0 && owner != t.unitsOnTile[0].owner) {
-					owner.EnsureRelationshipExists(t.unitsOnTile[0].owner);
-				}
-				if (t.owningCity != null && owner != t.owningCity.owner) {
-					owner.EnsureRelationshipExists(t.owningCity.owner);
-				}
-			}
 		}
 
 		public enum Intent {
@@ -1025,159 +616,6 @@ namespace C7GameData {
 			return tile.HasCity() && tile.cityAtTile.owner == player;
 		}
 
-		/// <summary>
-		/// Moves the unit in the given direction
-		/// </summary>
-		/// <param name="unit"></param>
-		/// <param name="dir">Which direction to move, e.g. northeast, west, etc.</param>
-		/// <param name="wait">Whether the method should wait to return until animations complete</param>
-		/// <returns>True if the unit is alive after the movement, false otherwise</returns>
-		/// <exception cref="Exception"></exception>
-		public async Task<bool> move(TileDirection dir, bool wait = false) {
-			(int dx, int dy) = dir.toCoordDiff();
-
-			Tile newLoc = EngineStorage.gameData.map.tileAt(dx + location.XCoordinate, dy + location.YCoordinate);
-
-			var canMove = (newLoc != Tile.NONE) && this.CanEnter(newLoc) && (movementPoints.canMove);
-			if (!canMove) return false;
-
-			facingDirection = dir;
-			wake();
-
-			// Trigger combat if the tile we're moving into has an enemy  Or if this unit can't fight, do nothing.
-			MapUnit defender = newLoc.FindTopDefender(this);
-			if (defender != MapUnit.NONE && !owner.IsAtPeaceWith(defender.owner)) {
-				if (unitType.attack <= 0) {
-					return true;
-				}
-
-				CombatResult combatResult = await fight(defender);
-				this.path = TilePath.NONE;
-				// If we were killed then of course there's nothing more to do. If the combat couldn't happen for whatever
-				// reason, just give up on trying to move.
-				if (combatResult == CombatResult.AttackerKilled) {
-					return false;
-				}
-				if (combatResult == CombatResult.Impossible) {
-					return true;
-				}
-
-				// If the enemy was defeated, check if there is another enemy on the tile. If so we can't complete the move
-				// but still pay one movement point for the combat.
-				if (combatResult == CombatResult.DefenderKilled || combatResult == CombatResult.DefenderRetreated) {
-					this.movementPoints.onUnitMove(1);
-					if (newLoc.FindTopDefender(this) != MapUnit.NONE) {
-						this.facingDirection = this.facingDirection.reversed();
-						return true;
-					}
-
-					// Similarly if we retreated, pay one MP for the combat but don't move.
-				} else if (combatResult == CombatResult.AttackerRetreated) {
-					this.movementPoints.onUnitMove(1);
-					this.facingDirection = this.facingDirection.reversed();
-					return true;
-				}
-			}
-
-			facingDirection = dir;
-			float movementCost = TilePath.GetMovementCost(this.owner, location, dir, newLoc);
-
-			// Leave old tile
-			if (!location.unitsOnTile.Remove(this))
-				throw new System.Exception("Failed to remove unit from tile it's supposed to be on");
-
-			// Move transported units, too
-			if (CanTransport()) {
-				var transported = location.unitsOnTile
-					.Where(u => u.IsLoadedIn(this)).ToList();
-
-				foreach (var tu in transported) {
-					if (!location.unitsOnTile.Remove(tu))
-						throw new System.Exception("Failed to remove unit from tile during transport move");
-					newLoc.unitsOnTile.Add(tu);
-					tu.location = newLoc;
-				}
-			}
-
-			TryBoardingTransportOnTile(newLoc);
-			TryUnboardingTransportToTile(newLoc);
-
-			// Enter new tile
-			// Make sure the unit is on the new location before claiming we have entered the tile
-			newLoc.unitsOnTile.Add(this);
-			location = newLoc;
-			OnEnterTile(newLoc);
-
-			if (wait)
-				await animateAsync(MapUnit.AnimatedAction.RUN);
-			else
-				animate(MapUnit.AnimatedAction.RUN);
-
-			movementPoints.onUnitMove(movementCost);
-
-			return true;
-		}
-
-		public void TryBoardingTransportOnTile(Tile newLoc) {
-			var enteringCity = newLoc.HasCity() && newLoc != location;
-			if (enteringCity || !CanBoardTransportOnTile(newLoc))
-				return;
-
-			var t = SelectTransportToBoard(newLoc);
-			BoardTransport(t);
-		}
-
-		public void TryUnboardingTransportToTile(Tile newLoc) {
-			if (!CanUnboardTransportToTile(newLoc))
-				return;
-
-			var t = FindTransportToUnboard(this.location, this.loadedOnUnitId);
-			UnboardTransport(t);
-		}
-
-		public void BoardTransport(MapUnit t) {
-			if (t == null) {
-				// TODO: throw new System.Exception("Failed to find a transport to move to");
-				Log.Warning("Failed to find a transport to board");
-				return;
-			}
-			t.board(this);
-			isFortified = true;
-			ResetFacingDirection();
-			if (this.owner.isHuman)
-				new MsgUnitMoved(this).send();
-		}
-
-		public void UnboardTransport(MapUnit t) {
-			if (t == null) {
-				// TODO: throw new System.Exception("Failed to find the transport to unboard from");
-				Log.Warning("Failed to find a transport to unboard");
-				return;
-			}
-			t.unboard(this);
-			wake();
-			if (this.owner.isHuman)
-				new MsgUnitMoved(this).send();
-		}
-
-		/// <summary>
-		/// Boards unit into this transport
-		/// </summary>
-		/// <param name="mapUnit">The unit to load on a transport</param>
-		private void board(MapUnit mapUnit) {
-			mapUnit.loadedOnUnitId = this.id;
-			// TODO: consume moves?
-		}
-
-		/// <summary>
-		/// Unloads a unit from this transport
-		/// </summary>
-		/// <param name="mapUnit">The unit to unload from a transport</param>
-		private void unboard(MapUnit mapUnit) {
-			mapUnit.loadedOnUnitId = null;
-			// TODO: consume moves?
-		}
-
 		private float SumWorkerProgress(Tile tile, Terraform workerJob) {
 			float result = 0;
 			foreach (MapUnit unit in tile.unitsOnTile) {
@@ -1205,63 +643,6 @@ namespace C7GameData {
 			return (int)Math.Ceiling(remainingTerraformCost / combinedWorkerSpeed);
 		}
 
-		public async Task PerformBusyAction() {
-			if (isFortified) {
-				return;
-			}
-
-			if (path != null && path.PathLength() > 0) {
-				await moveAlongPath();
-				return;
-			}
-
-			if (isAutomated) {
-				// workers contribute their work at the end of the turn, not when assigned
-				if (this.unitType.isWorker && WorkerJob != null) {
-					return;
-				}
-				playAutomatedTurn();
-				return;
-			}
-		}
-
-		public async Task PerformEndOfTurnAction() {
-			// Busy Worker
-			if (WorkerJob != null) {
-				WorkerProgressTowardsJob += workerSpeed();
-				movementPoints.onConsumeAll();
-
-				// See if this worker finished the job.
-				if ((int)SumWorkerProgress(location, WorkerJob) >= GetWorkerJobCost(location, WorkerJob)) {
-					location.FinishWorkerJob(WorkerJob);
-				}
-			}
-		}
-
-		public async Task moveAlongPath() {
-			while (movementPoints.canMove && path?.PathLength() > 0) {
-				TileDirection dir = location.directionTo(path.Next());
-				await move(dir, true); //TODO: don't wait on last move animation?
-			}
-		}
-
-		public async Task setUnitPath(TilePath path) {
-			this.path = path;
-			await moveAlongPath();
-		}
-
-		public void skipTurn() {
-			movementPoints.skipTurn();
-		}
-
-		public async Task Disband() {
-			await EngineStorage.gameData.DisbandUnit(this);
-		}
-
-		public void RemoveFromPlay() {
-			EngineStorage.gameData.RemoveUnit(this);
-		}
-
 		public bool canBuildCity() {
 			if (!unitType.actions.Contains(UnitAction.BuildCity)) {
 				return false;
@@ -1270,22 +651,6 @@ namespace C7GameData {
 				return false;
 			}
 			return location.neighbors.Values.All(tile => !tile.HasCity());
-		}
-
-		public async Task<City?> buildCity(string cityName) {
-			if (!canBuildCity()) {
-				log.Warning($"can't build city at {location}");
-				return null;
-			}
-
-			await animateAsync(MapUnit.AnimatedAction.BUILD);
-
-			// TODO: Need to check somewhere that this unit is allowed to build a city on its current tile. Either do that here or in every caller
-			// (probably best to just do it here).
-			City city = CityInteractions.BuildCity(location, owner, cityName);
-			this.RemoveFromPlay();
-
-			return city;
 		}
 
 		public bool CanPerformTerraformAction(Terraform terraform) {
@@ -1299,32 +664,6 @@ namespace C7GameData {
 			return containsTerraform && meetsRequirements && !hasCity;
 		}
 
-		// entry point for "manual" job assignment
-		public void PerformTerraformAction(Terraform terraform) {
-			if (!CanPerformTerraformAction(terraform)) {
-				log.Warning($"can't perform {terraform.Name} by {this}");
-				return;
-			}
-			WorkerJob = terraform;
-
-			if (terraform.Animation is AnimatedAction animation)
-				animate(animation, AnimationEnding.Repeat);
-
-			movementPoints.onConsumeAll();
-
-			// See if this worker finished the job.
-			var terraformProgress = this.SumWorkerProgress(this.location, this.WorkerJob);
-			var turnProgress = this.location.GetCurrentUnaccountedJobProgress(terraform);
-			var totalCost = (float)GetWorkerJobCost(this.location, this.WorkerJob);
-
-			if (terraformProgress + turnProgress == totalCost) {
-				location.FinishWorkerJob(WorkerJob);
-			}
-
-			wake();
-			_ = PerformBusyAction();
-		}
-
 		public float workerSpeed() {
 			float progressPerTurn = this.IsCaptive() ? JOB_PROGRESS_SLAVE : JOB_PROGRESS_WORKER;
 			if (owner.civilization.traits.Contains(Civilization.Trait.Industrious)) {
@@ -1336,62 +675,15 @@ namespace C7GameData {
 		public void resetWorkerJob() {
 			WorkerJob = null;
 			WorkerProgressTowardsJob = 0;
-			animate(MapUnit.AnimatedAction.BLANK, AnimationEnding.Repeat);
+			animate(AnimatedAction.BLANK, AnimationEnding.Repeat);
 		}
 
 		public bool canAutomate() {
 			return unitType.actions.Contains(UnitAction.Automate);
 		}
 
-		public void automate() {
-			wake();
-			isAutomated = true;
-			WorkerAIData? maybeAiData = WorkerAI.MakeAiData(this, owner);
-			if (maybeAiData == null) {
-				log.Information($"Could not find anything to automate for {this} owned by {owner}");
-				isAutomated = false;
-				return;
-			}
-			currentAI = new WorkerAI(maybeAiData);
-			playAutomatedTurn();
-		}
-
 		public bool canExplore() {
 			return unitType.actions.Contains(UnitAction.Explore);
-		}
-
-		public void explore() {
-			wake();
-			isAutomated = true;
-			ExplorerAIData? maybeAiData = ExplorerAI.MaybeMakeAiData(this, owner);
-			if (maybeAiData == null) {
-				log.Information($"Could not find anything to explore for {this} owned by {owner}");
-				isAutomated = false;
-				return;
-			}
-			currentAI = new ExplorerAI(maybeAiData);
-			playAutomatedTurn();
-		}
-
-		public async void playAutomatedTurn() {
-			if (currentAI == null) {
-				// TODO: handle giving automated workers from loaded saves the
-				// proper unit ai.
-				isAutomated = false;
-				return;
-			}
-			UnitAI.Result result = await currentAI.PlayTurn(owner, this);
-			if (result == UnitAI.Result.Done) {
-				if (currentAI is WorkerAI) {
-					automate();
-				} else if (currentAI is ExplorerAI) {
-					explore();
-				}
-			}
-
-			// Do nothing after an error so control returns to the player, and
-			// nothing after an progress result, so that next turn continues the
-			// AI action.
 		}
 
 		public List<Terraform> GetAvailableTerraforms() {
@@ -1435,28 +727,6 @@ namespace C7GameData {
 			// unit.availableActions.Add("rename");
 
 			return result;
-		}
-	}
-
-	public struct TileProbe {
-		public bool AllowCombat { get; init; }
-		public bool AllowWarDeclaration { get; init; }
-		public bool RaiseNotice { get; init; }
-
-		public static TileProbe MoveNonAggroProbe() {
-			return new TileProbe();
-		}
-
-		public static TileProbe MoveAggroProbe() {
-			return new TileProbe() { AllowCombat = true };
-		}
-
-		public static TileProbe MoveAggroWithNoticeProbe() {
-			return new TileProbe() { AllowCombat = true, RaiseNotice = true };
-		}
-
-		public static TileProbe DeclareWarProbe() {
-			return new TileProbe() { AllowCombat = true, AllowWarDeclaration = true };
 		}
 	}
 }
